requirements.txt
black==24.8.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
lightning[extra]==2.4.0
loguru==0.7.2
pytest==8.3.3
rich==13.8.1
rootutils==1.0.7
timm==1.0.9
pandas==2.2.3
numpy==1.26.0
torch==2.4.1
torchvision==0.19.1
flake8==7.1.1
dvc-gdrive==3.0.1
dvc-azure==3.1.0
transformers==4.45.2
pydantic-settings
uvicorn==0.32.0
tenacity==9.0.0
gunicorn==23.0.0
aim==3.25.0
mlflow==2.17.1
hydra-optuna-sweeper==1.2.0
dvc==3.56.0
platformdirs==3.10
fastapi-utils==0.7.0
httpx==0.27.2
typing-inspect==0.9.0
requests==2.32.3
fastapi-restful[all]==0.6.0
aioredis==2.0.1
psycopg2-binary==2.9.10
asyncpg==0.30.0
azure-servicebus==7.12.3
aiohttp==3.10.10
aiofiles
aiologger==0.7.0
pyyaml==6.0.2
sqlalchemy-utils==0.41.2
sqlalchemy==2.0.36
alembic==1.13.3
fastapi-limiter==0.1.6
redis==5.0.8
redisearch==2.0.0
python-multipart
python-dotenv==1.0.1
celery==5.4.0
fastapi-cache2==0.2.2
aiocache==0.12.3
dvc-s3==3.2.0
litserve==0.2.4
gpustat==1.1.1
nvitop==1.3.2
pyopenssl==23.0.0
cryptography
accelerate==1.1.1
aws-cdk-lib==2.172.0
constructs
pydantic
python-fasthtml==0.10.1
shad4fast==1.3.1